int64 type fix

+++ sqlitestore.go
@@ -119,14 +119,14 @@ func (s *SQLiteStore) GetLastBlock(ctx context.Context) (uint64, error) {
 }
 
 type blockStats struct {
-	creates      int
-	createsBytes int
-	updates      int
-	updatesBytes int
-	deletes      int
-	deletesBytes int
-	extends      int
-	ownerChanges int
+	creates      int64
+	createsBytes int64
+	updates      int64
+	updatesBytes int64
+	deletes      int64
+	deletesBytes int64
+	extends      int64
+	ownerChanges int64
 }
 
 func (s *SQLiteStore) FollowEvents(ctx context.Context, iterator arkivevents.BatchIterator) error {
@@ -199,7 +199,7 @@ func (s *SQLiteStore) FollowEvents(ctx context.Context, iterator arkivevents.Bat
 					case operation.Create != nil:
 						// expiresAtBlock := blockNumber + operation.Create.BTL
 						blockStat.creates++
-						blockStat.createsBytes += len(operation.Create.Content)
+						blockStat.createsBytes += int64(len(operation.Create.Content))
 						key := operation.Create.Key
 
 						stringAttributes := maps.Clone(operation.Create.StringAttributes)
@@ -262,7 +262,7 @@ func (s *SQLiteStore) FollowEvents(ctx context.Context, iterator arkivevents.Bat
 							continue operationLoop
 						}
 						blockStat.updates++
-						blockStat.updatesBytes += len(operation.Update.Content)
+						blockStat.updatesBytes += int64(len(operation.Update.Content))
 
 						key := operation.Update.Key.Bytes()
 
@@ -361,7 +361,7 @@ func (s *SQLiteStore) FollowEvents(ctx context.Context, iterator arkivevents.Bat
 						if err != nil {
 							return fmt.Errorf("failed to get latest payload: %w", err)
 						}
-						blockStat.deletesBytes += len(latestPayload.Payload)
+						blockStat.deletesBytes += int64(len(latestPayload.Payload))
 
 						oldStringAttributes := latestPayload.StringAttributes
 
@@ -501,14 +501,14 @@ func (s *SQLiteStore) FollowEvents(ctx context.Context, iterator arkivevents.Bat
 
 			// Calculate batch totals for logging and update metrics PER BLOCK
 			var (
-				totalCreates      int
-				totalCreatesBytes int
-				totalUpdates      int
-				totalUpdatesBytes int
-				totalDeletes      int
-				totalDeletesBytes int
-				totalExtends      int
-				totalOwnerChanges int
+				totalCreates      int64
+				totalCreatesBytes int64
+				totalUpdates      int64
+				totalUpdatesBytes int64
+				totalDeletes      int64
+				totalDeletesBytes int64
+				totalExtends      int64
+				totalOwnerChanges int64
 			)
 
 			// Iterate blocks again to preserve order and update metrics per block
